refactor: frontend folder structure
still wip
