+++ frontend/src/auth/components/FortyTwoIntraSignInButton.tsx
@@ -1,4 +1,4 @@
-import styles from '../../styles/Login.module.css';
+import styles from './styles/Login.module.css';
 import { useRouter } from 'next/router';
 
 const FortyTwoIntraSignInButton = () => {
