+++ frontend/src/chat/components/CatButton.tsx
