+++ frontend/src/chat/components/MessageInput.tsx
