+++ frontend/src/chat/components/ParticipatingChannelList.tsx
