+++ frontend/src/chat/components/SendMessageButton.tsx
