+++ frontend/src/game/conponents/End.tsx
