+++ frontend/src/game/conponents/ScoreBoard.tsx
