+++ frontend/src/game/conponents/Start.tsx
