--- frontend/src/styles/Common.module.css
@@ -1,7 +1,0 @@
-.commonLayout {
-  height: 100vh;
-  position: absolute;
-  right: 20%;
-  left: 20%;
-}
-
